+++ lib/widgets/webview/webview_notice_handlers.dart
@@ -10,6 +10,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_inappwebview/flutter_inappwebview.dart';
 import 'package:url_launcher/url_launcher.dart';
 
+import 'captcha_webview_dialog.dart';
 import 'download_options.dart';
 
 /// JavaScript handlers and download logic for WebViewNoticePage.
@@ -49,12 +50,7 @@ mixin WebViewNoticeHandlers<T extends StatefulWidget> on State<T> {
     final task = manager.taskFor(url, options.attachmentDir);
     if (task == null) return;
 
-    await _downloadWithCaptchaHandling(
-      url,
-      task.fileName,
-      options,
-      task,
-    );
+    await _downloadWithCaptchaHandling(url, task.fileName, options, task);
   }
 
   void onOpenImage(List<dynamic> args) {
@@ -164,7 +160,7 @@ mixin WebViewNoticeHandlers<T extends StatefulWidget> on State<T> {
     final result = await showDialog<bool>(
       context: context,
       barrierDismissible: false,
-      builder: (ctx) => _CaptchaWebViewDialog(
+      builder: (ctx) => CaptchaWebViewDialog(
         url: url,
         onDownloadComplete: (String filePath) {
           getIt<DownloadManager>().updateTask(
@@ -221,7 +217,10 @@ mixin WebViewNoticeHandlers<T extends StatefulWidget> on State<T> {
       url,
       options.attachmentDir,
       name,
-      headers: mergeDownloadHeaders(options.downloadHeaders, cookieHeader: null),
+      headers: mergeDownloadHeaders(
+        options.downloadHeaders,
+        cookieHeader: null,
+      ),
     );
 
     try {
@@ -277,99 +276,6 @@ mixin WebViewNoticeHandlers<T extends StatefulWidget> on State<T> {
     if (downloadOptions == null) return null;
     return DownloadStartResponse(
       handled: await handleDownloadStartRequest(request),
-    );
-  }
-}
-
-// ── CAPTCHA WebView Dialog ────────────────────────────────────────────────────
-
-/// Dialog that loads the CAPTCHA page directly in a WebView.
-///
-/// The user completes the verification form on the original server page. When
-/// the server returns the actual file after verification, [onDownloadStarting]
-/// intercepts it, downloads the file, and calls [onDownloadComplete].
-class _CaptchaWebViewDialog extends StatefulWidget {
-  const _CaptchaWebViewDialog({
-    required this.url,
-    required this.onDownloadComplete,
-    required this.getCookies,
-    required this.attachmentDir,
-    this.downloadHeaders,
-  });
-
-  final String url;
-  final void Function(String filePath) onDownloadComplete;
-  final Future<String?> Function() getCookies;
-  final String attachmentDir;
-  final Map<String, String>? downloadHeaders;
-
-  @override
-  State<_CaptchaWebViewDialog> createState() => _CaptchaWebViewDialogState();
-}
-
-class _CaptchaWebViewDialogState extends State<_CaptchaWebViewDialog> {
-  bool _loading = true;
-
-  Future<DownloadStartResponse?> _onDownloadStarting(
-    InAppWebViewController controller,
-    DownloadStartRequest request,
-  ) async {
-    final cookieHeader = await widget.getCookies();
-    final headers = mergeDownloadHeaders(
-      widget.downloadHeaders,
-      cookieHeader: cookieHeader,
-    );
-    try {
-      final path = await downloadFile(
-        request.url.toString(),
-        widget.attachmentDir,
-        request.suggestedFilename ?? 'download',
-        headers: headers,
-      );
-      widget.onDownloadComplete(path);
-    } catch (_) {
-      // Download failed — let the WebView keep showing whatever the server returned.
-    }
-    return DownloadStartResponse(handled: true);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Dialog(
-      child: SizedBox(
-        width: 420,
-        height: 260,
-        child: Column(
-          children: [
-            AppBar(
-              title: Text('请输入验证码'),
-              leading: IconButton(
-                icon: const Icon(Icons.close),
-                onPressed: () => Navigator.pop(context, false),
-              ),
-              toolbarHeight: 44,
-            ),
-            Expanded(
-              child: Stack(
-                children: [
-                  InAppWebView(
-                    initialUrlRequest: URLRequest(url: WebUri(widget.url)),
-                    initialSettings: InAppWebViewSettings(
-                      javaScriptEnabled: true,
-                    ),
-                    onDownloadStarting: _onDownloadStarting,
-                    onLoadStop: (ctrl, uri) {
-                      if (mounted) setState(() => _loading = false);
-                    },
-                  ),
-                  if (_loading)
-                    const Center(child: CircularProgressIndicator()),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
     );
   }
 }
